Move chart to be reused and fix decimal errors in input
